requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
botocore==1.13.33
cachetools==3.1.1
certifi==2019.11.28
chalice==1.12.0
chardet==3.0.4
Click==7.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
flake8==3.7.9
Flask==1.1.1
gast==0.2.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.750
mypy-extensions==0.4.3
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pep8==1.7.1
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
python-dateutil==2.8.0
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
six==1.13.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
